Refactor test suites.
