+++ main/test/ca/uwaterloo/flix/language/LanguageSuite.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2016 Magnus Madsen
+ * Copyright 2016 Magnus Madsen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,13 @@
  * limitations under the License.
  */
 
-package ca.uwaterloo.flix
+package ca.uwaterloo.flix.language
 
-import ca.uwaterloo.flix.language.LanguageSuite
-import ca.uwaterloo.flix.runtime.RuntimeSuite
-import ca.uwaterloo.flix.util.UtilSuite
-
+import ca.uwaterloo.flix.language.phase.PhaseSuite
 import org.scalatest.{ParallelTestExecution, Suites}
 
-class TestAll extends Suites(
-  new LanguageSuite,
-  new RuntimeSuite,
-  new UtilSuite,
-  new TestMain,
-  new TestExamples
+class LanguageSuite extends Suites(
+  new PhaseSuite
 ) with ParallelTestExecution {
   /* left empty */
 }
