+++ main/test/ca/uwaterloo/flix/language/phase/PhaseSuite.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2016 Magnus Madsen
+ * Copyright 2016 Magnus Madsen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,15 @@
  * limitations under the License.
  */
 
-package ca.uwaterloo.flix
-
-import ca.uwaterloo.flix.language.LanguageSuite
-import ca.uwaterloo.flix.runtime.RuntimeSuite
-import ca.uwaterloo.flix.util.UtilSuite
+package ca.uwaterloo.flix.language.phase
 
 import org.scalatest.{ParallelTestExecution, Suites}
 
-class TestAll extends Suites(
-  new LanguageSuite,
-  new RuntimeSuite,
-  new UtilSuite,
-  new TestMain,
-  new TestExamples
+class PhaseSuite extends Suites(
+  new TestParser,
+  new TestResolver,
+  new TestTyper,
+  new TestWeeder
 ) with ParallelTestExecution {
   /* left empty */
 }
