+++ main/test/ca/uwaterloo/flix/util/UtilSuite.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2016 Magnus Madsen
+ * Copyright 2016 Magnus Madsen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,12 @@
  * limitations under the License.
  */
 
-package ca.uwaterloo.flix
-
-import ca.uwaterloo.flix.language.LanguageSuite
-import ca.uwaterloo.flix.runtime.RuntimeSuite
-import ca.uwaterloo.flix.util.UtilSuite
+package ca.uwaterloo.flix.util
 
 import org.scalatest.{ParallelTestExecution, Suites}
 
-class TestAll extends Suites(
-  new LanguageSuite,
-  new RuntimeSuite,
-  new UtilSuite,
-  new TestMain,
-  new TestExamples
+class UtilSuite extends Suites(
+  new TestValidation
 ) with ParallelTestExecution {
   /* left empty */
 }
